A0 is a 10 x 10 matrix representing the upper triangle of a actuarial development triangle used by Roger Hayne to illustrate the stochastic loss reserving software in

A0 is a 10 x 10 matrix representing the upper triangle of a actuarial development triangle used by Roger Hayne to illustrate the stochastic loss reserving software in

Usage

A0

Format

A 10 x 10 numeric matrix

A0

Development triangle.

Source

"A Flexible Framework for Stochastic Reserving Models" found at https://variancejournal.org/article/120823