Get the long description of a model

Get the long description of a model

Usage

model_description(model)

Arguments

model

character vector of length one having the short name of the model.

Value

Long description of a model as a character vector of length one.